+++ CHANGELOG.md
@@ -35,7 +35,7 @@
 - Improved validation on Account parameter on `*-CosmosDBUserDefinedFunction*` functions.
 - Improved validation on Database parameter on `*-CosmosDBUserDefinedFunction*` functions.
 - Improved validation on Collection parameter on `*-CosmosDBUserDefinedFunction*` functions.
-- Improved validation on User Defined Function Id parameter on
+- Improved validation on User Id parameter on
  `*-CosmosDBUser*` functions - fixes [Issue #217](https://github.com/PlagueHO/CosmosDB/issues/217).
 - Improved validation on Account parameter on `*-CosmosDBUser*` functions.
 - Improved validation on Database parameter on `*-CosmosDBUser*` functions.
@@ -44,6 +44,11 @@
 - Improved validation on Account parameter on `*-CosmosDBDocument*` functions.
 - Improved validation on Database parameter on `*-CosmosDBDocument*` functions.
 - Improved validation on Collection parameter on `*-CosmosDBDocument*` functions.
+- Improved validation on Permission Id parameter on
+ `*-CosmosDBPermission*` functions - fixes [Issue #218](https://github.com/PlagueHO/CosmosDB/issues/218).
+- Improved validation on Account parameter on `*-CosmosDBPermission*` functions.
+- Improved validation on Database parameter on `*-CosmosDBPermission*` functions.
+- Improved validation on User parameter on `*-CosmosDBPermission*` functions.
 
 ## 2.1.12.137